Update bilara-io
Now bilara-io lives in its own folder.
A subfolder "config" is used to store config for import.
Added command line arguments -c/--create for creating new files
and --config for manually specifying a config file for creation.
