+++ .scripts/pyproject.toml
@@ -15,6 +15,7 @@ numpy = "^1.20.1"
 biopython = "^1.78"
 pip = "^21.0.1"
 inflection = "^0.5.1"
+pyexcel = "^0.6.6"
 
 [tool.poetry.dev-dependencies]
 black = {version = "^20.8b1", allow-prereleases = true}
